+++ include/depthai/pipeline/datatype/NNData.hpp
@@ -141,7 +141,7 @@ class NNData : public Buffer {
     std::vector<std::int32_t> getFirstLayerInt32() const;
 
     /**
-     * Retrieves image timestamp related to steady_clock / time.monotonic
+     * Retrieves image timestamp related to dai::Clock::now()
      */
     std::chrono::time_point<std::chrono::steady_clock, std::chrono::steady_clock::duration> getTimestamp() const;
 
@@ -154,7 +154,7 @@ class NNData : public Buffer {
     /**
      * Retrieves image sequence number
      */
-    unsigned int getSequenceNum() const;
+    int64_t getSequenceNum() const;
 };
 
 }  // namespace dai
