+++ include/depthai/pipeline/datatype/SpatialImgDetections.hpp
@@ -30,7 +30,7 @@ class SpatialImgDetections : public Buffer {
     std::vector<SpatialImgDetection>& detections;
 
     /**
-     * Retrieves image timestamp related to steady_clock / time.monotonic
+     * Retrieves image timestamp related to dai::Clock::now()
      */
     std::chrono::time_point<std::chrono::steady_clock, std::chrono::steady_clock::duration> getTimestamp() const;
 
@@ -43,7 +43,7 @@ class SpatialImgDetections : public Buffer {
     /**
      * Retrieves image sequence number
      */
-    unsigned int getSequenceNum() const;
+    int64_t getSequenceNum() const;
 };
 
 }  // namespace dai
